cleanup some mess I made while updating the TCP to http. Once again supports multiple nodes on the same computer
