--- lib/test.html
@@ -1,8 +1,0 @@
-<html>
-  <body>
-    <p id="main"></p>
-  </body>
-  <script>
-
-  </script>
-</html>
